requirements.txt
amqp==1.4.6
anyjson==0.3.3
beautifulsoup4==4.3.2
billiard==3.3.0.19
celery==3.1.17
Django==1.6.5
django-angular==0.7.5
django-autocomplete-light==1.4.14
django-celery==3.1.16
django-celery-admin==0.1
django-filter==0.7
django-rest-swagger==0.1.14
django-windows-tools==0.1.1
djangorestframework==2.3.14
ecdsa==0.13
Fabric==1.8.2
flup==1.0.2
httpretty==0.8.0
kombu==3.0.24
lxml==3.3.6
mock==1.0.1
paramiko==1.12.2
pycrypto==2.6.1
pytz==2014.10
requests==2.2.1
six==1.7.3
South==1.0.2
urllib3==1.10.4
virtualenv==15.1.0
